fully implemented tdx verification
